Move range_format and format_kind to their own header range_format.h
